refactor: split storage module to meet 400-line limit The TemporalDb storage module exceeded the 400-line constraint in both the implementation file (567 lines) and its test file (553 lines). Split strategy: - storage_types.rs — row types (HotspotRow, RiskRow, CochangeRow) - storage_ops.rs — store/load/sync impl block for TemporalDb - storage.rs — struct definition, open(), migrations, schema_version() re-exports row types so the public API (lib.rs) is unchanged - storage_tests.rs — groups 1-5: schema, CRUD, meta (322 lines) - storage_perf_tests.rs — groups 6-8: persistence, performance, errors (258 lines) All 31 storage tests continue to pass. No public API changes. Co-Authored-By: Claude <noreply@anthropic.com>
